+++ server/src/main/java/com/costa/socket/server/ServerApplication.java
@@ -1,0 +1,9 @@
+package com.costa.socket.server;
+
+import com.costa.socket.server.service.ServerRunner;
+
+public class ServerApplication {
+    public static void main(String[] args) {
+        new ServerRunner().start();
+    }
+}
