+++ server/src/main/java/com/costa/socket/server/dao/UserDao.java
@@ -1,0 +1,11 @@
+package com.costa.socket.server.dao;
+
+import com.costa.socket.server.model.User;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserDao {
+    Optional<User> findByName(String name);
+    List<User> findAll();
+}
