+++ server/src/main/java/com/costa/socket/server/exception/MapperException.java
@@ -1,0 +1,15 @@
+package com.costa.socket.server.exception;
+
+public class MapperException extends Exception{
+    public MapperException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public MapperException(String message) {
+        super(message);
+    }
+
+    public MapperException() {
+        super();
+    }
+}
