+++ server/src/main/java/com/costa/socket/server/model/ResultStatus.java
@@ -1,0 +1,12 @@
+package com.costa.socket.server.model;
+
+public enum ResultStatus {
+    CREATED,
+    DELETED,
+    CLOSED,
+    REOPENED,
+    TASKS,
+    WRONG_FORMAT,
+    ACCESS_DENIED,
+    ERROR
+}
