+++ server/src/main/java/com/costa/socket/server/model/TaskCommand.java
@@ -1,0 +1,9 @@
+package com.costa.socket.server.model;
+
+public enum TaskCommand {
+    CREATE_TASK,
+    DELETE_TASK,
+    CLOSE_TASK,
+    REOPEN_TASK,
+    LIST_TASK
+}
