+++ server/src/main/java/com/costa/socket/server/model/TaskState.java
@@ -1,0 +1,5 @@
+package com.costa.socket.server.model;
+
+public enum TaskState {
+    CREATED, CLOSED, DELETED
+}
