+++ server/src/main/java/com/costa/socket/server/service/CommandHandler.java
@@ -1,0 +1,7 @@
+package com.costa.socket.server.service;
+
+import com.costa.socket.server.dto.ServerResponse;
+
+public interface CommandHandler {
+    ServerResponse handle(String command);
+}
