+++ server/src/main/java/com/costa/socket/server/service/CommandValidator.java
@@ -1,0 +1,6 @@
+package com.costa.socket.server.service;
+
+public interface CommandValidator {
+    boolean validateComposition(String command);
+    boolean validateAccess(String command);
+}
