--- server/src/main/java/com/garipov/socket/server/Server1.java
@@ -1,51 +1,0 @@
-package com.garipov.socket.server;
-
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
-import java.net.ServerSocket;
-import java.net.Socket;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class Server1 {
-
-  private static final Logger LOG = LoggerFactory.getLogger(Server1.class);
-
-  private ServerSocket serverSocket;
-
-  public void start() throws IOException {
-    serverSocket = new ServerSocket(9090);
-    Thread serverThread = new Thread(() -> {
-      while (true) {
-        try {
-          Socket connection = serverSocket.accept();
-          try (
-              BufferedReader serverReader = new BufferedReader(
-                  new InputStreamReader(connection.getInputStream()));
-              Writer serverWriter = new BufferedWriter(
-                  new OutputStreamWriter(connection.getOutputStream()))
-          ) {
-            String line = serverReader.readLine();
-            LOG.debug("Request captured: {}", line);
-            // В реализации по умолчанию в ответе пишется та же строка, которая пришла в запросе
-            serverWriter.write(line);
-            serverWriter.flush();
-          }
-        } catch (Exception e) {
-          LOG.error("Error during request proceeding", e);
-          break;
-        }
-      }
-    });
-    serverThread.setDaemon(true);
-    serverThread.start();
-  }
-
-  public void stop() throws IOException {
-    serverSocket.close();
-  }
-}
