--- server/src/main/java/com/garipov/socket/server/ServerApplication.java
@@ -1,9 +1,0 @@
-package com.garipov.socket.server;
-
-import com.garipov.socket.server.service.ServerRunner;
-
-public class ServerApplication {
-    public static void main(String[] args) {
-        new ServerRunner().start();
-    }
-}
