--- server/src/main/java/com/garipov/socket/server/dto/UserRequest.java
@@ -1,36 +1,0 @@
-package com.garipov.socket.server.dto;
-
-import com.garipov.socket.server.model.TaskCommand;
-
-public class UserRequest {
-    private final String userName;
-    private final TaskCommand command;
-    private final String arg;
-
-    public UserRequest(String userName, TaskCommand command, String arg) {
-        this.userName = userName;
-        this.command = command;
-        this.arg = arg;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public TaskCommand getCommand() {
-        return command;
-    }
-
-    public String getArg() {
-        return arg;
-    }
-
-    @Override
-    public String toString() {
-        return "UserRequest{" +
-                "userName='" + userName + '\'' +
-                ", command=" + command +
-                ", arg='" + arg + '\'' +
-                '}';
-    }
-}
