--- server/src/main/java/com/garipov/socket/server/model/ResultStatus.java
@@ -1,12 +1,0 @@
-package com.garipov.socket.server.model;
-
-public enum ResultStatus {
-    CREATED,
-    DELETED,
-    CLOSED,
-    REOPENED,
-    TASKS,
-    WRONG_FORMAT,
-    ACCESS_DENIED,
-    ERROR
-}
