--- server/src/main/java/com/garipov/socket/server/model/TaskCommand.java
@@ -1,9 +1,0 @@
-package com.garipov.socket.server.model;
-
-public enum TaskCommand {
-    CREATE_TASK,
-    DELETE_TASK,
-    CLOSE_TASK,
-    REOPEN_TASK,
-    LIST_TASK
-}
