--- server/src/main/java/com/garipov/socket/server/model/TaskState.java
@@ -1,5 +1,0 @@
-package com.garipov.socket.server.model;
-
-public enum TaskState {
-    CREATED, CLOSED, DELETED
-}
