--- server/src/main/java/com/garipov/socket/server/service/CommandHandler.java
@@ -1,7 +1,0 @@
-package com.garipov.socket.server.service;
-
-import com.garipov.socket.server.dto.ServerResponse;
-
-public interface CommandHandler {
-    ServerResponse handle(String command);
-}
