--- server/src/main/java/com/garipov/socket/server/service/CommandValidator.java
@@ -1,6 +1,0 @@
-package com.garipov.socket.server.service;
-
-public interface CommandValidator {
-    boolean validateComposition(String command);
-    boolean validateAccess(String command);
-}
